rename file _gateway
